MidiMusic.deltaTimeSpec is now Int. Porting MidiPlayerTest from managed-midi.
The new test is not fully implemented as it needs file resolver which needs
some implementation in MPP, or we dump MPP for testing.

+++ src/jvmMain/kotlin/dev/atsushieno/ktmidi/MidiReaderWriter.kt
@@ -250,7 +250,7 @@ class SmfReader(private var stream: InputStream) {
             throw parseError("Unexpected data size (should be 6)")
         data.format = readInt16().toByte()
         val tracks = readInt16()
-        data.deltaTimeSpec = readInt16().toByte()
+        data.deltaTimeSpec = readInt16().toInt()
         for (i in 0 until tracks)
             data.tracks.add(readTrack())
     }
